Rename handle_map to handle_maps
